Create a Python S3 Get Contents
Python S3 Get Contents
In the Python file, write a program to access the contents of the bucket "coderbytechallengesandbox". In there, there might be multiple files, but your program should find the file with the prefix __cb__, and then output the contents of that file. You should use the boto3 module to solve this challenge.

You do not need any access keys to access the bucket because it is public. This post might help you with how to access the bucket.

Example Output
contents of some file
Browse Resources
Search for any help or documentation you might need for this problem. For example: an array indexing, Ruby hash tables, etc.